+++ R/CFAxisLongitude.R
@@ -47,8 +47,8 @@ CFAxisLongitude <- R6::R6Class("CFAxisLongitude",
     #' @return The newly created axis.
     copy = function(name = "") {
       if (self$has_resource) {
-        ax <- CFAxisLongitude$new(self$NC, values = self$values, start = private$.start_count$start,
-                                  count = private$.start_count$count, attributes = self$attributes)
+        ax <- CFAxisLongitude$new(self$NC, values = self$values, start = private$.NC_map$start,
+                                  count = private$.NC_map$count, attributes = self$attributes)
         if (nzchar(name))
           ax$name <- name
       } else {
@@ -93,7 +93,7 @@ CFAxisLongitude <- R6::R6Class("CFAxisLongitude",
       else {
         rng <- range(rng)
         if (self$has_resource) {
-          ax <- CFAxisLongitude$new(private$.NCobj, start = private$.start_count$start + rng[1L] - 1L,
+          ax <- CFAxisLongitude$new(private$.NCobj, start = private$.NC_map$start + rng[1L] - 1L,
                                     count = rng[2L] - rng[1L] + 1L, attributes = self$attributes)
           if (nzchar(name))
             ax$name <- name