+++ R/CFData.R
@@ -9,15 +9,21 @@ CFData <- R6::R6Class("CFData",
   inherit = CFObject,
   cloneable = FALSE,
   private = list(
-    # The values of this object. This class does not manipulate the values in
+    # The values of this object. This class does not interpret the values in
     # any way - that is the job of descendant classes.
     .values = NULL,
 
+    # Integer vector of the dimensions of .values. It is set at initialization
+    # on the basis of .values. This should include all dimensions of the owning
+    # object, including degenerate ones. The owning object should set this
+    # parameter with any degenerate dimensions.
+    .dims = integer(0),
+
     # List of start and count vectors for reading data from file. The length of
     # the vectors must agree with the array on file or be `NA`.
-    .start_count = list(start = NA, count = NA),
+    .NC_map = list(),
 
-    # The data type of the data in the object. Taken from .NCobj if set,
+    # The NC data type of the data in the object. Taken from .NCobj if set,
     # otherwise the descending class must set it explicitly when receiving its
     # data values.
     .data_type = "NC_NAT", # Not a type
@@ -25,30 +31,45 @@ CFData <- R6::R6Class("CFData",
     # Flag to indicate if the data on disk is too large for memory.
     .data_oversized = FALSE,
 
-    # Sanitize the start and count values. NAs are converted to numbers and
-    # values have to agree with .values and .NCobj. Returns the sanitized start
-    # and count vectors as a list.
+    # Set the dimensions of the data in this object. The product of `new_dims`
+    # must equal the actual .dims. This method is really only useful to add
+    # length-1 dimensions, such as a scalar axis, to the current dimensions.
+    set_dims = function(new_dims) {
+      if (prod(private$.dims) == prod(new_dims)) {
+        names(new_dims) <- NULL
+        private$.dims <- new_dims
+        if (!is.null(private$.values))
+          dim(private$.values) <- private$.dims
+      } else
+        stop("New dimensions do not agree with existing data dimensions.", call. = FALSE)
+    },
+
+    # Sanitize the start and count values. Called during initialization and by
+    # read_chunk(). NAs are converted to numbers and values have to agree with
+    # .dims. Returns the sanitized start and count vectors as a list.
     check_start_count = function(start, count) {
-      d <- as.integer(self$dim())
-      if (!length(d)) # When a NC variable does not have any data
-        return(list(start = NA, count = NA))
+      d <- private$.dims
+      len <- length(d)
+      if (!len) return(list())          # When the object does not have any data
 
       if (length(start) == 1L && is.na(start))
-        start <- rep(1L, length(d))
-      else
+        start <- rep(1L, len)
+      else if (length(start) == len) {
         start[is.na(start)] <- 1L
+        if (any(start > d))
+          stop("Start values cannot be larger than the dimensions of the data.", call. = FALSE) # nocov
+      } else
+        stop("`start` vector is not the length of the object dimensions.", call. = FALSE)
 
-      if (any(start > d))
-        stop("Start values cannot be larger than the dimensions of the data.", call. = FALSE) # nocov
       if (length(count) == 1L && is.na(count))
         count <- d - start + 1L
-      else {
+      else if (length(count) == len) {
         ndx <- which(is.na(count))
         count[ndx] <- d[ndx] - start[ndx] + 1L
-      }
-
-      if (any(count > d - start + 1L))
-        stop("Count values cannot extend beyond the dimensions of the data.", call. = FALSE) # nocov
+        if (any(count > d - start + 1L))
+          stop("Count values cannot extend beyond the dimensions of the data.", call. = FALSE) # nocov
+      } else
+        stop("`count` vector is not the length of the object dimensions.", call. = FALSE)
 
       list(start = start, count = count)
     },
@@ -56,7 +77,7 @@ CFData <- R6::R6Class("CFData",
     # Set the values of the object. Perform some basic checks when set programmatically.
     # Values may be NULL
     set_values = function(values) {
-      # FIXME: Check that dim(values) agrees with NCvar and data_type
+      # FIXME: Check that values agrees with .dims, .NC_map and .data_type
 
       # Check if we can find a vtype from the NCvar, possibly packed
       if (!is.null(private$.NCobj)) {
@@ -91,8 +112,8 @@ CFData <- R6::R6Class("CFData",
       if (is.null(values))
         self$delete_attribute("actual_range")
       else {
-        rng <- range(values, na.rm = TRUE)
-        if (is.na(rng[1L]))
+        rng <- suppressWarnings(range(values, na.rm = TRUE))
+        if (is.infinite(rng[1L]) || is.na(rng[1L]))
           self$delete_attribute("actual_range")
         else {
           if (is.numeric(rng))
@@ -101,15 +122,18 @@ CFData <- R6::R6Class("CFData",
         }
       }
 
+      if (!is.null(values) && self$ndims > 1L) {
+        dim(values) <- private$.dims
+      }
       private$.values <- values
     },
 
-    # Read the data of the CF object from file. The data is cached by `self` so
-    # repeated calls do not access the netCDF resource, unless argument
-    # `refresh` is `TRUE`.
-    # This method will not assess how big the data is before reading it so there
-    # is a chance that memory will be exhausted. The calling code should check
-    # for this possibility and break up the reading of data into chunks.
+    # Read the data of the CF object from the NC file. The data is cached by
+    # `self` so repeated calls do not access the netCDF resource, unless
+    # argument `refresh` is `TRUE`. This method will not assess how big the data
+    # is before reading it so there is a chance that memory will be exhausted.
+    # The calling code should check for this possibility and break up the
+    # reading of data into chunks.
     # @param refresh Should the data be read from file if the object is linked?
     #   This will replace current values, if previously loaded. Default `FALSE`.
     # @return An array of data, invisibly, as prescribed by the `start` and
@@ -117,69 +141,86 @@ CFData <- R6::R6Class("CFData",
     #   by a netCDF resource, returns `NULL`.
     read_data = function(refresh = FALSE) {
       if ((!is.null(private$.NCobj)) && (is.null(private$.values) || refresh))
-        private$set_values(private$.NCobj$get_data(private$.start_count$start, private$.start_count$count))
+        private$set_values(private$.NCobj$get_data(private$.NC_map$start, private$.NC_map$count))
       invisible(private$.values)
     },
 
-    # Read a chunk of data of the CF object, as defined by the `start` and
+    # Read a chunk of raw data of the CF object, as defined by the `start` and
     # `count` vectors. Note that these vectors are relative to any subset of the
-    # netCDF data variable that this CF object refers to. The data read by this
-    # method will not be stored in `self` so the calling code must take a
-    # reference to it.
+    # data variable that this CF object refers to. The data read by this method
+    # will not be stored in `self` so the calling code must take a reference to
+    # it.
     # @param start Vector of indices where to start reading data along the
     #   dimensions of the array. The vector must be `NA` to read all data,
     #   otherwise it must have agree with the dimensions of the array.
     # @param count Vector of number of elements to read along each dimension of
-    #   the array on file. The vector must be `NA` to read to the end of each
-    #   dimension, otherwise its value must agree with the corresponding `start`
-    #   value and the dimension of the array.
+    #   the array. The vector must be `NA` to read to the end of each dimension,
+    #   otherwise its value must agree with the corresponding `start` value and
+    #   the dimension of the array.
     # @return An array of data, as prescribed by the `start` and `count`
     #   arguments, or `NULL` if there is no data.
     read_chunk = function(start, count) {
       sc <- private$check_start_count(start, count)
-      if (is.na(sc$start[1L])) return(NULL)
+      if (!length(sc)) return (NULL)
 
       if (!is.null(private$.values)) {
         # Extract from loaded data
         cll <- paste0("private$.values[", paste(sc$start, ":", sc$start + sc$count - 1L, sep = "", collapse = ", "), "]")
         eval(parse(text = cll))
       } else {
-        # Read from the netCDF resource
-        start <- private$.start_count$start + sc$start - 1L
-        private$.NCobj$get_data(start, sc$count)
+        # Read from the netCDF resource. .NC_map always refers to the initial
+        # dimensions, so the arguments are trimmed to that length (noting that
+        # there may be "scalar" axes in a variable backed by a netCDF resource).
+        len <- length(private$.NC_map$start)
+        start <- private$.NC_map$start + sc$start[1L:len] - 1L
+        private$.NCobj$get_data(start, sc$count[1L:len])
       }
     }
   ),
   public = list(
-    #' @description Create a new `CFobject` instance from a variable in a netCDF
-    #'   resource. This method is called upon opening a netCDF resource. It is
-    #'   rarely, if ever, useful to call this constructor directly. Instead, use
-    #'   the methods from higher-level classes such as [CFVariable].
+    #' @description Create a new `CFData` instance from a variable in a netCDF
+    #'   resource. This method is called upon creating CF objects, such as when
+    #'   opening a netCDF resource or creating a new CF object. It is rarely, if
+    #'   ever, useful to call this constructor directly. Instead, use the
+    #'   methods from higher-level classes such as [CFVariable].
     #' @param obj The [NCVariable] instance upon which this CF object is based
     #'   when read from a netCDF resource, or the name for the new CF object to
     #'   be created.
-    #' @param values Optional. The values of the object in an array.
+    #' @param values Optional. The values of the object in an array. Ignored
+    #'   when argument `obj` is an `NCVariable` instance.
     #' @param start Optional. Vector of indices where to start reading data
     #'   along the dimensions of the array on file. The vector must be `NA` to
     #'   read all data, otherwise it must have agree with the dimensions of the
-    #'   array on file. Ignored when argument `obj` is not an `NCVariable`
+    #'   array on file. Default value is `1`, i.e. start from the beginning of
+    #'   the 1-D NC variable. Ignored when argument `obj` is not an `NCVariable`
     #'   instance.
     #' @param count Optional. Vector of number of elements to read along each
     #'   dimension of the array on file. The vector must be `NA` to read to the
     #'   end of each dimension, otherwise its value must agree with the
     #'   corresponding `start` value and the dimension of the array on file.
-    #'   Ignored when argument `obj` is not an `NCVariable` instance.
+    #'   Default is `NA`. Ignored when argument `obj` is not an `NCVariable`
+    #'   instance.
     #' @param attributes Optional. A `data.frame` with the attributes of the
     #'   object.
     #' @return A `CFData` instance.
     initialize = function(obj, values, start = 1L, count = NA, attributes = data.frame()) {
       super$initialize(obj, attributes)
 
-      if (inherits(obj, "NCObject"))
-        private$.start_count <- private$check_start_count(start, count)
+      if (inherits(obj, "NCVariable")) {
+        # Set the initial .NC_map to cover the entire NC variable
+        if ((nd <- obj$ndims) > 0L) {
+          private$.dims <- obj$dim()
+          private$.NC_map <- list(start = rep(1L, nd), count = private$.dims)
+        }
+        # Subset to a smaller dimension space, if needed
+        private$.NC_map <- private$check_start_count(start, count)
+        private$.dims <- private$.NC_map$count
+      }
 
-      if (missing(values) || (length(values) == 1L && is.na(values)))
+      if (missing(values) || is.null(values) || (length(values) == 1L && is.na(values)))
         values <- NULL
+      else
+        private$.dims <- dim(values) %||% length(values)
       private$set_values(values)
     },
 
@@ -192,34 +233,21 @@ CFData <- R6::R6Class("CFData",
       super$detach()
     },
 
-    #' @description Retrieve the dimensions of the data of this object. This
-    #'   could be for the data on file or for in-memory data.
+    #' @description Retrieve the dimensions of the data of this object.
     #' @param dimension Optional. The index of the dimension to retrieve the
     #'   length for. If omitted, retrieve the lengths of all dimensions.
     #' @return Integer vector with the length of each requested dimension.
     dim = function(dimension) {
-      len <- if (!is.null(private$.values)) {
-        d <- dim(private$.values)
-        if (length(d) == 0L) length(private$.values) else d
-      } else if (self$has_resource) {
-        if (length(private$.start_count$count) > 1L || !is.na(private$.start_count$count))
-          private$.start_count$count - private$.start_count$start + 1L
-        else
-          private$.NCobj$dim()
-      } else
-        return(NULL)
-
       if (missing(dimension))
-        len
+        private$.dims
       else
-        len[dimension]
+        private$.dims[dimension]
     }
-
   ),
   active = list(
-    #' @field data_type Set or retrieve the data type of the data in the
-    #'   object. Setting the data type to a wrong value can have unpredictable
-    #'   and mostly catastrophic consequences.
+    #' @field data_type Set or retrieve the data type of the data in the object.
+    #'   Setting the data type to a wrong value can have unpredictable and
+    #'   mostly catastrophic consequences.
     data_type = function(value) {
       if (missing(value))
         private$.data_type
@@ -232,18 +260,18 @@ CFData <- R6::R6Class("CFData",
     },
 
     #' @field ndims (read-only) Retrieve the dimensionality of the data in the
-    #'   array, or the netCDF resource.
+    #'   array.
     ndims = function(value) {
       if (missing(value))
-        length(self$dim())
+        length(private$.dims)
     },
 
-    #' @field array_indices Returns a list with columns "start" and "count"
-    #'   giving the indices for reading the data of this object from a netCDF
-    #'   resource.
-    array_indices = function(value) {
+    #' @field NC_map Returns a list with columns "start" and "count" giving the
+    #'   indices for reading the data of this object from a netCDF resource. The
+    #'   list is empty if this object is not backed by a netCDF resource.
+    NC_map = function(value) {
       if (missing(value))
-        private$.start_count
+        private$.NC_map
     }
   )
 )